+++ config/src/keyassignment.rs
@@ -130,7 +130,7 @@ impl Default for Pattern {
 }
 
 /// A mouse event that can trigger an action
-#[derive(Debug, Clone, PartialEq, Eq, Hash, FromDynamic, ToDynamic)]
+#[derive(Debug, Clone, PartialEq, Eq, Ord, PartialOrd, Hash, FromDynamic, ToDynamic)]
 pub enum MouseEventTrigger {
     /// Mouse button is pressed. streak is how many times in a row
     /// it was pressed.